api/requirements.txt
bcrypt==3.1.7
cffi==1.14.1
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
python-dotenv==0.14.0
six==1.15.0
Werkzeug==1.0.1
